Move ComponentMetricRequest structure to avoid cyclic imports
