+++ src/frequenz/sdk/actor/_data_sourcing/data_sourcing.py
@@ -7,7 +7,8 @@ from frequenz.channels import Receiver
 
 from .._actor import Actor
 from .._channel_registry import ChannelRegistry
-from .microgrid_api_source import ComponentMetricRequest, MicrogridApiSource
+from ._component_metric_request import ComponentMetricRequest
+from .microgrid_api_source import MicrogridApiSource
 
 
 class DataSourcingActor(Actor):
